Updated helper and interface
