+++ src/interfaces/ChatImagesResponse.ts
@@ -1,0 +1,9 @@
+import { MessageDetails } from './MessageDetails';
+
+export interface ChatImagesResponse {
+  chatId: string;
+  totalImages: number;
+  totalChats: number;
+  messagesWithImages: MessageDetails[];
+  skipToken?: string;
+}
